scamper-cvs-20190916

+++ scamper/scamper_options.h
@@ -1,7 +1,7 @@
 /*
  * scamper_options.h: code to handle parsing of options
  *
- * $Id: scamper_options.h,v 1.7 2012/04/05 18:00:54 mjl Exp $
+ * $Id: scamper_options.h,v 1.8 2019/07/12 23:37:57 mjl Exp $
  *
  * Copyright (C) 2006-2008 The University of Waikato
  * Author: Matthew Luckie
@@ -97,7 +97,8 @@ int scamper_options_parse(char *opt_str,
 
 int scamper_options_validate(const scamper_option_in_t *opts, const int cnt,
 			     int argc, char *argv[], int *stop,
-			     int validate(int optid, char *param, long *out));
+			     int validate(int optid, char *param,
+					  long long *out));
 
 /*
  * scamper_options_count
